HDFSLineageDispatcher and AbstractJsonLineageDispatcher uses runtime json4s dependencies in order to prevent failing on binary incompatibilities (e.g. between json4s 3.5.3 and 3.5.5)
